tests: Explicitly spawn a session libvirt-dbus instance
Tests are performed using the session D-Bus instance, so we
should launch libvirt-dbus in session mode as well.

This was working fine when running the tests as a regular
user, because in that case libvirt-dbus would default to
session mode, but fail when running them as root because
libvirt-dbus would run in system mode and consequently not
show up on the session bus.

Of course building and runnning the test suite as root is
a pretty bad idea in general, but a lot of distributions
run at least part of their package build steps with pretend
root privileges (eg. fakeroot), so we have to make sure it
works in that scenario too.

Signed-off-by: Andrea Bolognani <abologna@redhat.com>
Reviewed-by: Katerina Koukiou <kkoukiou@redhat.com>