tests: Make node_device_create() work with libvirt 3.0.0
The scsi_host2 nodedev, which our test nodedev uses as its
parent, was added to the test driver with commit 5c2ff641e10c,
which is included in libvirt 3.1.0; before that, there was a
similar device called test-scsi-host-vport, which is no longer
present after libvirt 3.0.0.

Since there's no nodedev that we can use as parent both with
libvirt 3.0.0 and with later versions, our only option is to
perform a lookup at runtime and use whatever's available.

Signed-off-by: Andrea Bolognani <abologna@redhat.com>
Reviewed-by: Katerina Koukiou <kkoukiou@redhat.com>