util: fix virtDBusUtilDecodeUUID
This function is supposed to convert ASCII character into its hex
representation, however the current implementation was wrong because
the first comparison would be false for all printable characters.  In
most cases it worked but for example '$' which is 0x24 in HEX would be
incorrectly converted to 0x2[ which is obviously wrong.

Resolves: https://bugzilla.redhat.com/show_bug.cgi?id=1647823

Signed-off-by: Pavel Hrdina <phrdina@redhat.com>