build: convert to Meson/Ninja build system
Meson build system is simple and quick compared to Autotools and it's
able to fully replace our Autotools usage.  There are few drawbacks as
it's a fairly new build system, it requires Python 3.5 and Ninja 1.5.0,
it's still evolving and the user base is not that large and there were
some tweaks required to achieve the same functionality.

However, there are benefits, the configure and build time is way shorter
and build definition files are more readable and easier to maintain.

There are some major changes with Meson build system:

    - there is no syntax-check target, the syntax-check is part of Meson
      test suite but it's still possible to run it separately,

    - Meson forces separation between source and build directories

Signed-off-by: Pavel Hrdina <phrdina@redhat.com>
Tested-by: Ján Tomko <jtomko@redhat.com>
Reviewed-by: Fabiano Fidêncio <fidencio@redhat.com>