meson: Install documentation
People who install the RPM packages already get this, but it's
good practice so let's make it happen for those who install from
source as well.

The process is straightforward, but we have to be a bit careful
with AUTHORS because this specific file is found in different
directories depending on whether we're building from git or from
a release.

Signed-off-by: Andrea Bolognani <abologna@redhat.com>
Reviewed-by: Pavel Hrdina <phrdina@redhat.com>