gitlab: introduce CI jobs testing git master & distro libvirt
The dbus build needs to validate one axis

  - A variety of libvirt versions

We test a variety of libvirt versions by running a build against the
distro provided libvirt packages. All that is then missing is a build
against the latest libvirt git master, which only needs to be run on
a single distro, for which CentOS 8 & Stream are picked as a stable
long life base.

RPM build and tests are skipped on CentOS 7 since it lacks the required
min version of some packages.

Reviewed-by: Andrea Bolognani <abologna@redhat.com>
Signed-off-by: Daniel P. Berrangé <berrange@redhat.com>